# api-service/openapi.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "api-service",
		"description": "guggr api definition",
		"contact": {
			"name": "Guggr"
		},
		"license": {
			"name": ""
		},
		"version": "0.1.0"
	},
	"paths": {
...
		"/api/v1/jobs": {
			"get": {
				"tags": ["jobs"],
				"summary": "list endpoint for jobs",
				"operationId": "list_job",
				"responses": {
					"200": {
						"description": "List jobs",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/DisplayJob"
									}
								}
							}
						}
					},
					"500": {
						"description": "Storage error"
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				]
			},
			"post": {
				"tags": ["jobs"],
				"summary": "create endpoint for jobs",
				"operationId": "create_job",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateJob"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "Created job",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DisplayJob"
								}
							}
						}
					},
					"500": {
						"description": "Storage error"
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				]
			}
		},
		"/api/v1/jobs/{id}": {
			"get": {
				"tags": ["jobs"],
				"summary": "get endpoint for jobs",
				"operationId": "get_job",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"description": "Job id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Job",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DisplayJob"
								}
							}
						}
					},
					"500": {
						"description": "Storage error"
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				]
			},
			"delete": {
				"tags": ["jobs"],
				"summary": "delete endpoint for jobs",
				"operationId": "delete_job",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"description": "Job id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"204": {
						"description": "Deleted"
					},
					"500": {
						"description": "Storage error"
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				]
			},
			"patch": {
				"tags": ["jobs"],
				"summary": "update endpoint for jobs",
				"operationId": "update_job",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateJob"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "Created job",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DisplayJob"
								}
							}
						}
					},
					"500": {
						"description": "Storage error"
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				]
			}
		},
		"/api/v1/jobs/{id}/runs": {
			"get": {
				"tags": ["jobs"],
				"summary": "list endpoint for job runs",
				"operationId": "list_job_runs",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"description": "Job id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "List jobs runs",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/DisplayJobRun"
									}
								}
							}
						}
					},
					"500": {
						"description": "Storage error"
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				]
			}
		},
...
		"/api/v1/roles": {
			"get": {
				"tags": ["roles"],
				"summary": "list endpoint for roles",
				"operationId": "list_role",
				"responses": {
					"200": {
						"description": "List roles",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/DisplayRole"
									}
								}
							}
						}
					},
					"500": {
						"description": "Storage error"
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				]
			},
			"post": {
				"tags": ["roles"],
				"summary": "create endpoint for roles",
				"operationId": "create_role",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateRole"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "Created role",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DisplayRole"
								}
							}
						}
					},
					"400": {
						"description": "Validation error"
					},
					"500": {
						"description": "Storage error"
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				]
			}
		},
		"/api/v1/roles/{id}": {
			"get": {
				"tags": ["roles"],
				"summary": "get endpoint for roles",
				"operationId": "get_role",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"description": "Role id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Role",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DisplayRole"
								}
							}
						}
					},
					"500": {
						"description": "Storage error"
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				]
			},
			"delete": {
				"tags": ["roles"],
				"summary": "delete endpoint for roles",
				"operationId": "delete_role",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"description": "Role id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"204": {
						"description": "Deleted"
					},
					"500": {
						"description": "Storage error"
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				]
			},
			"patch": {
				"tags": ["roles"],
				"summary": "update endpoint for roles",
				"operationId": "update_role",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"description": "Role id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateRole"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "Patched role",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DisplayRole"
								}
							}
						}
					},
					"400": {
						"description": "Validation error"
					},
					"500": {
						"description": "Storage error"
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				]
			}
		},
...
	},
	"components": {
		"schemas": {
...
			"CreateJob": {
				"type": "object",
				"description": "sent to the database for creating a new job",
				"required": ["name", "job_type_id", "group_id", "notify_users", "run_every", "details"],
				"properties": {
					"custom_notification": {
						"type": ["string", "null"]
					},
					"details": {
						"$ref": "#/components/schemas/CreateJobDetails"
					},
					"group_id": {
						"type": "string"
					},
					"job_type_id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"notify_users": {
						"type": "boolean"
					},
					"run_every": {
						"type": "string"
					}
				}
			},
			"CreateJobDetails": {
				"oneOf": [
					{
						"type": "object",
						"required": ["Http"],
						"properties": {
							"Http": {
								"$ref": "#/components/schemas/CreateJobDetailsHttp"
							}
						}
					},
					{
						"type": "object",
						"required": ["Ping"],
						"properties": {
							"Ping": {
								"$ref": "#/components/schemas/CreateJobDetailsPing"
							}
						}
					}
				],
				"description": "sent to the database for creating a new job and its details"
			},
			"CreateJobDetailsHttp": {
				"type": "object",
				"description": "sent to the database for creating a new job detail",
				"required": ["url"],
				"properties": {
					"url": {
						"type": "string"
					}
				}
			},
			"CreateJobDetailsPing": {
				"type": "object",
				"description": "sent to the database for creating a new job detail",
				"required": ["host"],
				"properties": {
					"host": {
						"type": "string"
					}
				}
			},
			"CreateRole": {
				"type": "object",
				"description": "sent to the database for creating a new role",
				"required": ["name"],
				"properties": {
					"name": {
						"type": "string",
						"minLength": 1
					}
				}
			},
...
			"DisplayJob": {
				"type": "object",
				"description": "returned from the database when displaying a job",
				"required": [
					"id",
					"name",
					"job_type_id",
					"group_id",
					"notify_users",
					"run_every",
					"details"
				],
				"properties": {
					"custom_notification": {
						"type": ["string", "null"]
					},
					"details": {
						"$ref": "#/components/schemas/DisplayJobDetails"
					},
					"group_id": {
						"type": "string"
					},
					"id": {
						"type": "string"
					},
					"job_type_id": {
						"type": "string"
					},
					"last_scheduled": {
						"type": ["string", "null"],
						"format": "date-time"
					},
					"name": {
						"type": "string"
					},
					"notify_users": {
						"type": "boolean"
					},
					"run_every": {
						"type": "string"
					}
				}
			},
			"DisplayJobDetails": {
				"oneOf": [
					{
						"type": "object",
						"required": ["Http"],
						"properties": {
							"Http": {
								"$ref": "#/components/schemas/DisplayJobDetailsHttp"
							}
						}
					},
					{
						"type": "object",
						"required": ["Ping"],
						"properties": {
							"Ping": {
								"$ref": "#/components/schemas/DisplayJobDetailsPing"
							}
						}
					}
				],
				"description": "returned from the database when displaying a job and its details"
			},
			"DisplayJobDetailsHttp": {
				"type": "object",
				"description": "returned from the database when displaying a job detail",
				"required": ["id", "url"],
				"properties": {
					"id": {
						"type": "string"
					},
					"url": {
						"type": "string"
					}
				}
			},
			"DisplayJobDetailsPing": {
				"type": "object",
				"description": "returned from the database when displaying a job detail",
				"required": ["id", "host"],
				"properties": {
					"host": {
						"type": "string"
					},
					"id": {
						"type": "string"
					}
				}
			},
			"DisplayJobResultHttp": {
				"type": "object",
				"description": "returned from the database when displaying a job result",
				"required": ["id", "ip_address", "status_code", "latency", "payload"],
				"properties": {
					"id": {
						"type": "string"
					},
					"ip_address": {
						"type": "string",
						"format": "cidr",
						"example": "192.0.2.5/32"
					},
					"latency": {
						"type": "integer",
						"format": "int32"
					},
					"payload": {
						"type": "array",
						"items": {
							"type": "integer",
							"format": "int32",
							"minimum": 0
						}
					},
					"status_code": {
						"type": "integer",
						"format": "int32"
					}
				}
			},
			"DisplayJobResultPing": {
				"type": "object",
				"description": "returned from the database when displaying a job result",
				"required": ["id", "ip_address", "latency"],
				"properties": {
					"id": {
						"type": "string"
					},
					"ip_address": {
						"type": "string",
						"format": "cidr",
						"example": "192.0.2.5/32"
					},
					"latency": {
						"type": "integer",
						"format": "int32"
					}
				}
			},
			"DisplayJobRun": {
				"type": "object",
				"description": "returned from the database when displaying a jobrun",
				"required": [
					"id",
					"job_id",
					"timestamp",
					"triggered_notification",
					"batch_id",
					"reachable",
					"details"
				],
				"properties": {
					"batch_id": {
						"type": "string"
					},
					"details": {
						"$ref": "#/components/schemas/DisplayJobRunDetails"
					},
					"id": {
						"type": "string"
					},
					"job_id": {
						"type": "string"
					},
					"reachable": {
						"type": "boolean"
					},
					"timestamp": {
						"type": "string",
						"format": "date-time"
					},
					"triggered_notification": {
						"type": "boolean"
					}
				}
			},
			"DisplayJobRunDetails": {
				"oneOf": [
					{
						"type": "object",
						"required": ["Http"],
						"properties": {
							"Http": {
								"$ref": "#/components/schemas/DisplayJobResultHttp"
							}
						}
					},
					{
						"type": "object",
						"required": ["Ping"],
						"properties": {
							"Ping": {
								"$ref": "#/components/schemas/DisplayJobResultPing"
							}
						}
					}
				],
				"description": "returned from the database when displaying a jobrun with its details"
			},
			"DisplayRole": {
				"type": "object",
				"description": "returned from the database when displaying a role",
				"required": ["id", "name"],
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				}
			},
...
			"UpdateJob": {
				"type": "object",
				"description": "request by a user to update a job",
				"properties": {
					"custom_notification": {
						"type": ["string", "null"]
					},
					"details": {
						"oneOf": [
							{
								"type": "null"
							},
							{
								"$ref": "#/components/schemas/UpdateJobDetails"
							}
						]
					},
					"group_id": {
						"type": ["string", "null"]
					},
					"id": {
						"type": ["string", "null"]
					},
					"job_type_id": {
						"type": ["string", "null"]
					},
					"last_scheduled": {
						"type": ["string", "null"],
						"format": "date-time"
					},
					"name": {
						"type": ["string", "null"]
					},
					"notify_users": {
						"type": ["boolean", "null"]
					},
					"run_every": {
						"type": ["string", "null"]
					}
				}
			},
			"UpdateJobDetails": {
				"oneOf": [
					{
						"type": "object",
						"required": ["Http"],
						"properties": {
							"Http": {
								"$ref": "#/components/schemas/UpdateJobDetailsHttp"
							}
						}
					},
					{
						"type": "object",
						"required": ["Ping"],
						"properties": {
							"Ping": {
								"$ref": "#/components/schemas/UpdateJobDetailsPing"
							}
						}
					}
				],
				"description": "request by a user to update a job and its details"
			},
			"UpdateJobDetailsHttp": {
				"type": "object",
				"properties": {
					"url": {
						"type": ["string", "null"]
					}
				}
			},
			"UpdateJobDetailsPing": {
				"type": "object",
				"properties": {
					"host": {
						"type": ["string", "null"]
					}
				}
			},
			"UpdateRole": {
				"type": "object",
				"description": "sent to the database when updating a role",
				"properties": {
					"name": {
						"type": ["string", "null"],
						"minLength": 1
					}
				}
			},
...
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	}
}