# api-service/src/core/models/job/ping/result.rs
use frunk::LabelledGeneric;
use ipnet::IpNet;
use serde::{Deserialize, Serialize};
use utoipa::ToSchema;

#[derive(
    Debug, PartialEq, Eq, Clone, Deserialize, Serialize, ToSchema, Default, LabelledGeneric,
)]
/// returned from the database when displaying a job result
pub struct DisplayJobResultPing {
    pub id: String,
    #[schema(value_type = String, format = "cidr", example = "192.0.2.5/32")]
    pub ip_address: IpNet,
    pub latency: i32,
}